requirements.txt
Flask==2.1.2
Jinja2==3.1.2

python-dateutil==2.8.2
python-ldap==3.4.0

boto3==1.23.9
botocore==1.26.10
